move unittest into related test class
